Move useful setup helpers into tmtest
